Elimino soundpool, solo queda audiotrack

+++ app/src/main/java/com/mnlgt/elmetronomo/ui/MetronomoUiState.kt
@@ -1,7 +1,7 @@
 package com.mnlgt.elmetronomo.ui
 
 data class MetronomoUiState(
-    var bpm: Int = 60,
+    var bpm: Float = 60F,
     var andando: Boolean = false
 
 )
